requirements.txt
pypdf
langchain
langchain-community
chromadb # Vector storage
supabase
python-dotenv
pytest
boto3
torch
transformers
huggingface-hub
langchain-huggingface
sentence-transformers
